CListControl font support and better positioning
Add a proxy object to support font changes to list control.
Upon any change, a method is called which updates the font
and repositions the items and markers accordingly.

This repositioning has as well been refined and now replaces
the initial positoning. It takes care of automatic margins
and alignment.