fix loading of parts selected on TreeView
Remove old count of to-be-ignored events because it was
increased with every step within a part. This made it
impossible to change the part while some steps in the
current part had been made.

Instead, now check if the part to be loaded is already loaded
and ignore if so. That means it is no longer possible to
"restart" the current part, but that's unnecessary anyway.